test(input): migrate a11y
